Split RichContentFormatter out of WordPressSharedUI
RichContentFormatter is plain String -> String HTML sanitizing, but it lived in WordPressSharedUI, so WordPressData pulled in the UI layer to use it. Move the eight platform-independent methods and the RegEx table down to WordPressShared; keep formatContentString and resizeGalleryImageURL — which read UIScreen and call PhotonImageURLHelper — in WordPressSharedUI as an @objc extension. WordPressData no longer depends on WordPressSharedUI.

The text-formatting tests move with the code and now run on macOS under swift test; the gallery-resize test stays iOS-only.
