+++ Modules/Package.swift
@@ -318,7 +318,6 @@ let package = Package(
                 "FormattableContentKit",
                 "SFHFKeychainUtils",
                 "WordPressShared",
-                "WordPressSharedUI",
                 "WordPressKit",
                 .product(name: "CocoaLumberjack", package: "CocoaLumberjack"),
                 .product(name: "CocoaLumberjackSwift", package: "CocoaLumberjack"),
