+++ Modules/Tests/WordPressSharedTests/RichContentFormatterUITests.swift
@@ -1,0 +1,10 @@
+import XCTest
+@testable import WordPressShared
+@testable import WordPressSharedUI
+
+class RichContentFormatterUITests: XCTestCase {
+
+    func testResizeGalleryImageURLsForContentEmptyString() {
+        XCTAssertTrue("" == RichContentFormatter.resizeGalleryImageURL("", isPrivateSite: false))
+    }
+}
